contrib/fuel_bootstrap/fuel_bootstrap_cli/requirements.txt
PyYAML>=3.1.0
argparse
stevedore
pbr>=0.6
cliff>=1.7.0
six>=1.7.0
python-fuelclient
